Fix cidr_mask::str not correctly displaying mask length
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@12363 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/socket.cpp
@@ -331,7 +331,7 @@ std::string irc::sockets::cidr_mask::str() const
 	else
 		return "";
 	memcpy(base, bits, len);
-	return sa.addr() + "/" + ConvToStr(length);
+	return sa.addr() + "/" + ConvToStr((int)length);
 }
 
 bool irc::sockets::cidr_mask::operator==(const cidr_mask& other) const
